[FE] FEAT: banned user 해제 기능 구현 #985
