+++ frontend/src/hooks/useMenu.ts
@@ -98,6 +98,7 @@ const useMenu = () => {
   };
 
   const openCabinet = () => {
+    console.log("hi");
     closeLeftNav();
     closeMap();
     closeLent();
